+++ packages/auth-apple/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @zemble/auth-apple
 
+## 0.0.53
+
+### Patch Changes
+
+- c4186d0: Allow overriding sub in @zemble/auth-apple
+- Updated dependencies [c4186d0]
+  - @zemble/auth@0.0.50
+
+## 0.0.52
+
+### Patch Changes
+
+- Updated dependencies [acdedbd]
+- Updated dependencies [d17b6f3]
+  - @zemble/auth@0.0.49
+  - @zemble/core@0.1.16
+  - @zemble/graphql@0.0.48
+
 ## 0.0.51
 
 ### Patch Changes