+++ packages/auth-otp/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @zemble/auth-otp
 
+## 0.0.51
+
+### Patch Changes
+
+- c4186d0: Allow overriding sub in @zemble/auth-apple
+- Updated dependencies [c4186d0]
+  - @zemble/auth@0.0.50
+
+## 0.0.50
+
+### Patch Changes
+
+- acdedbd: Login with phone number
+- Updated dependencies [acdedbd]
+- Updated dependencies [d17b6f3]
+  - @zemble/auth@0.0.49
+  - @zemble/core@0.1.16
+  - @zemble/utils@0.7.35
+  - @zemble/graphql@0.0.48
+  - @zemble/kv@0.0.46
+
 ## 0.0.49
 
 ### Patch Changes