+++ packages/cms/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @zemble/cms
 
+## 0.0.55
+
+### Patch Changes
+
+- Updated dependencies [c4186d0]
+  - @zemble/auth@0.0.50
+
+## 0.0.54
+
+### Patch Changes
+
+- acdedbd: Login with phone number
+- Updated dependencies [acdedbd]
+- Updated dependencies [d17b6f3]
+  - @zemble/auth@0.0.49
+  - @zemble/core@0.1.16
+  - @zemble/mongodb@0.0.29
+  - @zemble/graphql@0.0.48
+
 ## 0.0.53
 
 ### Patch Changes