+++ packages/create-zemble-plugin/CHANGELOG.md
@@ -1,10 +1,17 @@
 # create-zemble-plugin
 
+## 0.1.1
+
+### Patch Changes
+
+- d17b6f3: Add @zemble/push-expo and support provider strategies
+
 ## 0.1.0
 
 ### Minor Changes
 
 - 1928354: Move in eslint package, rename graphql-codegen to codegen, other improvements.
+
   ```
 
   ```