+++ packages/graphql/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @zemble/graphql
 
+## 0.0.48
+
+### Patch Changes
+
+- Updated dependencies [d17b6f3]
+  - @zemble/core@0.1.16
+
 ## 0.0.47
 
 ### Patch Changes