+++ packages/node/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @zemble/node
 
+## 0.0.27
+
+### Patch Changes
+
+- Updated dependencies [d17b6f3]
+  - @zemble/core@0.1.16
+
 ## 0.0.26
 
 ### Patch Changes